ALPHA!
The darkVulkanGen application now generates a vulkan.pas file from the vk.xml which compiles and contains code for loading pointers for the api exports.
The loader depends on the darkDynLib library.
None of the functions have been tested to work, and compilation of vulkan.pas has only been tested on windows.