Manually applied conditional defines for nested types.
This is a bit of a hack. Manually added eight types to the conditional defines list because they are nested dependencies of other conditional elements. Writing a recursive algorithm to take care of them is a lot of effort to go to for the sake of eight types, so I simply added them manually. This may break things in the future, however, a major refactor of the XML parser is required anyway, so taking the easy way out pending a rewrite if necessary.